pyproject.toml
[tool.poetry]
name = "cq-source-productboard"
version = "0.1.0"
description = ""
authors = ["Pilvikala"]
readme = "README.md"
packages = [{include = "plugin"}]

[tool.poetry.dependencies]
python = ">=3.11"
cloudquery-plugin-sdk = "0.1.47"
pydantic = "^2.4"
pyarrow = "^19.0.1"
requests = "^2.32.2"

[tool.poetry.scripts]
main = "main:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

